compiler/next: fix parser EOF bug, add warnings, and fix warnings

This PR fixes a core dump when parsing something like `/* comment ` (an 
unterminated comment). It wasn't clear to me how to communicate to flex
that EOF had been reached with a push parser so this PR simply adds an 
`atEOF` bool to `ParserContext` and checks it while parsing.

While there, I noticed an unused variable that was not generating a 
compilation warning. So this PR also adjusts
`compiler/next/Makefile.help` to include the warning flags from the usual
Makefiles so that we get these. I fixed a number of warnings that came up
as a result.

To resolve a `clang` error with redundant dependency flags,
avoids adding the `DEPEND_CFLAGS` to the cmake configure step.

Reviewed by @dlongnecke-cray - thanks!

+++ compiler/next/lib/uast/FnCall.cpp
@@ -35,7 +35,7 @@ bool FnCall::contentsMatchInner(const ASTNode* other) const {
       lhs->actualNames_.size() != rhs->actualNames_.size())
     return false;
 
-  size_t nActualNames = lhs->actualNames_.size();
+  int nActualNames = (int) lhs->actualNames_.size();
   for (int i = 0; i < nActualNames; i++) {
     if (lhs->actualNames_[i] != rhs->actualNames_[i])
       return false;
